PaimonsNotebook 1.4.3

+++ app/src/main/java/com/lianyi/paimonsnotebook/common/application/PaimonsNotebookApplication.kt
@@ -8,14 +8,30 @@ import android.content.pm.ApplicationInfo
 import android.content.pm.ShortcutManager
 import android.os.Bundle
 import androidx.lifecycle.ProcessLifecycleOwner
+import cat.ereza.customactivityoncrash.config.CaocConfig
+import coil.ImageLoader
+import coil.ImageLoaderFactory
+import coil.decode.ImageDecoderDecoder
+import coil.disk.DiskCache
 import com.lianyi.paimonsnotebook.BuildConfig
 import com.lianyi.paimonsnotebook.R
 import com.lianyi.paimonsnotebook.common.database.PaimonsNotebookDatabase
+import com.lianyi.paimonsnotebook.common.util.builder.imageLoader
+import com.lianyi.paimonsnotebook.common.util.coil.MergeInterceptor
+import com.lianyi.paimonsnotebook.common.util.data_store.PreferenceKeys
+import com.lianyi.paimonsnotebook.common.util.data_store.dataStoreValuesFirstLambda
 import com.lianyi.paimonsnotebook.common.util.image.PaimonsNotebookImageLoader
+import com.lianyi.paimonsnotebook.common.util.request.emptyOkHttpClient
+import com.lianyi.paimonsnotebook.common.view.CrashScreen
+import com.lianyi.paimonsnotebook.ui.screen.splash.view.SplashScreen
+import com.microsoft.appcenter.AppCenter
+import com.microsoft.appcenter.analytics.Analytics
+import com.microsoft.appcenter.crashes.Crashes
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.first
 
-class PaimonsNotebookApplication : Application()  {
+
+class PaimonsNotebookApplication : Application(), ImageLoaderFactory {
     companion object {
         @SuppressLint("StaticFieldLeak")
         private lateinit var mContext: Context
@@ -23,9 +39,7 @@ class PaimonsNotebookApplication : Application()  {
             mContext
         }
 
-        val version by lazy {
-            BuildConfig.VERSION_NAME
-        }
+        const val version = BuildConfig.VERSION_NAME
 
         val name by lazy {
             context.getString(R.string.app_name)
@@ -44,15 +58,31 @@ class PaimonsNotebookApplication : Application()  {
         }
     }
 
-
     override fun onCreate() {
         super.onCreate()
         mContext = applicationContext
 
+        CaocConfig.Builder.create()
+            .backgroundMode(CaocConfig.BACKGROUND_MODE_SILENT)
+            .enabled(true)
+            .showErrorDetails(true)
+            .showRestartButton(false)
+            .logErrorOnRestart(false)
+            .minTimeBetweenCrashesMs(300)
+            .restartActivity(SplashScreen::class.java)
+            .errorActivity(CrashScreen::class.java)
+            .apply()
+
+        AppCenter.start(
+            this,
+            BuildConfig.APPCENTER_SECRET,
+            Analytics::class.java,
+            Crashes::class.java
+        )
+
         ProcessLifecycleOwner.get().lifecycle.addObserver(ApplicationLifecycleObserver())
 
 //        initActivityLifecycleCallbacks()
-
 //        initShortcutManager()
 
         executeDiskCachePlanDelete()
@@ -76,7 +106,7 @@ class PaimonsNotebookApplication : Application()  {
     //监听全部activity的状态
     private fun initActivityLifecycleCallbacks() {
 
-        registerActivityLifecycleCallbacks(object :ActivityLifecycleCallbacks{
+        registerActivityLifecycleCallbacks(object : ActivityLifecycleCallbacks {
             override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) {
                 println("activity:${activity.componentName.className} onActivityCreated")
             }
@@ -108,19 +138,53 @@ class PaimonsNotebookApplication : Application()  {
 
     }
 
+    //最后调用的时间戳与当前时间戳相差大于此数,则判断为无用图片,默认为7天(604800000L)
+    private val deleteTimeStampLimit = 604800000L
+
     //清除计划删除图片文件
     private fun executeDiskCachePlanDelete() {
+        //TODO 目前已知删除disckCache文件会导致删除图片再次缓存图片,再从本地读取缓存为null
         CoroutineScope(Dispatchers.IO).launch {
+            val autoClean = dataStoreValuesFirstLambda {
+                this[PreferenceKeys.EnableAutoCleanExpiredImages] ?: true
+            }
+
+            if (!autoClean) return@launch
+
             PaimonsNotebookDatabase.database.diskCacheDao.apply {
+                updateAllDataPlanDeleteStatus(System.currentTimeMillis(), deleteTimeStampLimit)
 
                 getPlanDeleteData().first().forEach {
                     PaimonsNotebookImageLoader.getCacheImageFileByUrl(it.url)?.delete()
                     PaimonsNotebookImageLoader.getCacheImageMetadataFileByUrl(it.url)?.delete()
                 }
 
-                removePlanDelete()
+                removeAllPlanDeleteData()
             }
         }
     }
 
+    private val imageCachePath by lazy {
+        context.filesDir.resolve("image_cache")
+    }
+
+    private val imageCache by lazy {
+        DiskCache.Builder()
+            .directory(imageCachePath)
+            .maxSizePercent(1.0)
+            .build()
+    }
+
+    override fun newImageLoader(): ImageLoader =
+        imageLoader {
+            components {
+                callFactory(emptyOkHttpClient)
+                add { result, options, _ -> ImageDecoderDecoder(result.source, options, false) }
+                add(MergeInterceptor)
+            }
+            diskCache(imageCache)
+            crossfade(300)
+            error(R.drawable.ic_image_error)
+            respectCacheHeaders(false)
+        }
 }
